docker-requirements.txt
pandas==1.5.2
prefect-gcp[cloud_storage]==0.3.0
protobuf==4.21.11
pyarrow==10.0.1
pandas-gbq==0.18.1
requests==2.31.0
pydantic==1.10.11
python-dotenv==1.0.0